+++ src/perf/java/com/github/ilyavy/perfomance/BookAnalysisBenchmark.java
@@ -5,19 +5,19 @@ import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import com.github.ilyavy.parser.BookTextParser;
-import com.github.ilyavy.parser.ConcurrentMapWithAtomicWordCountersUsingForkJoinBookTextParser;
-import com.github.ilyavy.parser.ConcurrentMapWithAtomicWordCountersUsingThreadsAndPhaser;
-import com.github.ilyavy.parser.ForkJoinBookTextParser;
-import com.github.ilyavy.parser.FuturesBookTextParser;
-import com.github.ilyavy.parser.ParallelStreamsBookTextParser;
-import com.github.ilyavy.parser.ParallelStreamsGroupingByBookTextParser;
-import com.github.ilyavy.parser.ParallelStreamsNaiveBookTextParser;
-import com.github.ilyavy.parser.ParallelStreamsToMapBookTextParser;
-import com.github.ilyavy.parser.ParallelStreamsWithOnlyFlatmapBookTextParser;
-import com.github.ilyavy.parser.SimpleSequentialBookTextParser;
-import com.github.ilyavy.parser.SingleStreamNaiveBookTextParser;
-import com.github.ilyavy.reader.BookFileReader;
+import com.github.ilyavy.service.parser.BookTextParser;
+import com.github.ilyavy.service.parser.ConcurrentMapWithAtomicWordCountersUsingForkJoinBookTextParser;
+import com.github.ilyavy.service.parser.ConcurrentMapWithAtomicWordCountersUsingThreadsAndPhaser;
+import com.github.ilyavy.service.parser.ForkJoinBookTextParser;
+import com.github.ilyavy.service.parser.FuturesBookTextParser;
+import com.github.ilyavy.service.parser.ParallelStreamsBookTextParser;
+import com.github.ilyavy.service.parser.ParallelStreamsGroupingByBookTextParser;
+import com.github.ilyavy.service.parser.ParallelStreamsNaiveBookTextParser;
+import com.github.ilyavy.service.parser.ParallelStreamsToMapBookTextParser;
+import com.github.ilyavy.service.parser.ParallelStreamsWithOnlyFlatmapBookTextParser;
+import com.github.ilyavy.service.parser.SimpleSequentialBookTextParser;
+import com.github.ilyavy.service.parser.SingleStreamNaiveBookTextParser;
+import com.github.ilyavy.service.reader.BookFileReader;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
